requirements.txt
antlr4-python3-runtime==4.8
matplotlib==3.2.1
python-telegram-bot==12.7
